Refactor: Move checkTradeId and isTradeIdValid to TradeValidation
